Added a trigger.onEntityCollide trigger, which executes the usergen vc function it contains the name of whenever any entity hits any obstruction.
It sets the following variables:
event.tx and event.ty are the tile coodinates hit.
event.entity is the entity who did the hitting.
event.zone is the zone of the tile being struck, as long as it was a tile obstruction being hit.  Otherwise, -1.
event.entity_hit is entity index of the entity being hit, as long as it was an entity-on-entity collision.  Otherwise, -1.

git-svn-id: http://www.verge-rpg.com/svn/verge3/trunk@244 85f9bd53-ed1e-0410-abaf-ef5a3fdb13a8